+++ lib/src/interfaces/CommandCompletedEvent.ts
@@ -1,0 +1,5 @@
+export interface CommandCompletedEvent {
+  commandId: string;
+  completionTime: number;
+  params: any;
+}
