+++ lib/src/interfaces/LifecycleEvent.ts
@@ -1,0 +1,12 @@
+export interface LifecycleEvent {
+  type: LifecycleEventType;
+  componentId: string;
+  componentName: string;
+}
+
+export const enum LifecycleEventType {
+  ComponentDidMount = 'ComponentDidMount',
+  ComponentDidAppear = 'ComponentDidAppear',
+  ComponentDidDisappear = 'ComponentDidDisappear',
+  ComponentWillUnmount = 'ComponentWillUnmount'
+}
